Added url search to thread pool. Also added basic stats output
