--- data/data.json
@@ -1,1 +1,0 @@
-"[\n   {\n      \"label\": \"cat\",\n      \"src\": \"https://i.natgeofe.com/n/548467d8-c5f1-4551-9f58-6817a8d2c45e/NationalGeographic_2572187_3x4.jpg\",\n      \"filename\": \"cat0.jpg\"\n   },\n   {\n      \"label\": \"cat\",\n      \"src\": \"https://th-thumbnailer.cdn-si-edu.com/bgmkh2ypz03IkiRR50I-UMaqUQc=/1000x750/filters:no_upscale():focal(1061x707:1062x708)/https://tf-cmsv2-smithsonianmag-media.s3.amazonaws.com/filer_public/55/95/55958815-3a8a-4032-ac7a-ff8c8ec8898a/gettyimages-1067956982.jpg\",\n      \"filename\": \"cat1.jpg\"\n   },\n   {\n      \"label\": \"cat\",\n      \"src\": \"https://www.wfla.com/wp-content/uploads/sites/71/2023/05/GettyImages-1389862392.jpg?w=2560&h=1440&crop=1\",\n      \"filename\": \"cat2.jpg\"\n   },\n   {\n      \"label\": \"cat\",\n      \"src\": \"https://www.alleycat.org/wp-content/uploads/2019/03/FELV-cat.jpg\",\n      \"filename\": \"cat3.jpg\"\n   },\n   {\n      \"label\": \"cat\",\n      \"src\": \"https://cdn.britannica.com/34/235834-050-C5843610/two-different-breeds-of-cats-side-by-side-outdoors-in-the-garden.jpg\",\n      \"filename\": \"cat4.jpg\"\n   },\n   {\n      \"label\": \"cat\",\n      \"src\": \"https://media.4-paws.org/5/b/4/b/5b4b5a91dd9443fa1785ee7fca66850e06dcc7f9/VIER%20PFOTEN_2019-12-13_209-2890x2000-1920x1329.jpg\",\n      \"filename\": \"cat5.jpg\"\n   },\n   {\n      \"label\": \"dog\",\n      \"src\": \"https://hips.hearstapps.com/ghk.h-cdn.co/assets/16/08/gettyimages-530330473.jpg?crop=0.659xw:0.990xh;0.123xw,0.00779xh&resize=980:*\",\n      \"filename\": \"dog0.jpg\"\n   },\n   {\n      \"label\": \"dog\",\n      \"src\": \"https://www.southernliving.com/thmb/Rz-dYEhwq_82C5_Y9GLH2ZlEoYw=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/gettyimages-837898820-1-4deae142d4d0403dbb6cb542bfc56934.jpg\",\n      \"filename\": \"dog1.jpg\"\n   },\n   {\n      \"label\": \"dog\",\n      \"src\": \"https://thumbor.forbes.com/thumbor/fit-in/900x510/https://www.forbes.com/advisor/wp-content/uploads/2023/07/top-20-small-dog-breeds.jpeg.jpg\",\n      \"filename\": \"dog2.jpg\"\n   },\n   {\n      \"label\": \"dog\",\n      \"src\": \"https://cdn.britannica.com/92/212692-050-D53981F5/labradoodle-dog-stick-running-grass.jpg\",\n      \"filename\": \"dog3.jpg\"\n   },\n   {\n      \"label\": \"dog\",\n      \"src\": \"https://www.nylabone.com/-/media/project/oneweb/nylabone/images/dog101/activities-fun/10-great-small-dog-breeds/maltese-portrait.jpg\",\n      \"filename\": \"dog4.jpg\"\n   },\n   {\n      \"label\": \"dog\",\n      \"src\": \"https://i.natgeofe.com/n/5f35194b-af37-4f45-a14d-60925b280986/NationalGeographic_2731043_3x4.jpg\",\n      \"filename\": \"dog5.jpg\"\n   }\n]"
