+++ resources/env.py
@@ -3,11 +3,11 @@ PATH = "C:\\Users\\evan9\\OneDrive\\Desktop\\school\\Applied Machine Learning\\p
 IMAGE_DIR = "images\\"
 DATA_DIR = "data\\"
 MAX_IMAGES = 6
-CLASSES = ["cat", "dog"]
+CLASSES = ["otter", "fish"] #, "shark", "squid", "turtle", "eel", "jellyfish", "dolphin", "seahorse", "crab"]
 BROWSER_LOCATION = "C:\\Program Files\\Mozilla Firefox\\firefox.exe"
 DRIVER = "resources\\driver.exe"
 IMAGE_EXTENSION = ".jpg"
 #Output
-DEBUG = True #Do you want debug messages?
+DEBUG = False #Do you want debug messages?
 VERBOSE = True #Do you want messages about the program?
 ZIP = True #Do you want to compress the data and images?
