+++ src/ImageObject.py
@@ -1,8 +1,6 @@
-import requests
-import io
+import requests, io
 from PIL import Image
-import json
-from resources.env import DEBUG, VERBOSE, IMAGE_EXTENSION
+from resources.env import DEBUG, IMAGE_EXTENSION
 from resources.textColors import redText, greenText
 
 # ImageObject Class. Used for storing scraped web element data. 
@@ -34,7 +32,7 @@ class ImageObject :
         data = {
             "label" : str(self.label),
             "src" : str(self.src),
-            "filename" : str(self.filename) + IMAGE_EXTENSION
+            "filename" : str(self.filename)
         }
         return data
     def downloadImage(self, downloadPath) : #dowload image from src to filename
@@ -43,19 +41,11 @@ class ImageObject :
             if (DEBUG) : print(greenText("Successful Request: " + self.src))
             image_file = io.BytesIO(image_content)
             image = Image.open(image_file)
-            file_path = downloadPath + self.filename + IMAGE_EXTENSION
+            file_path = downloadPath + self.filename
 
             with open(file_path, "wb") as f:
                 image.save(f, "JPEG")
-                print(file_path)
             if (DEBUG) : print(greenText("Saved image: " + file_path))
         except Exception as e:
             if (DEBUG) : print(redText('FAILED -' + str(e)))
 
-# JSON Output Function
-def outputJSON(outPath, jsonFilename, jsonData) :
-    jsonObject = json.dumps(jsonData, indent=3)
-    filePath = outPath + jsonFilename + ".json"
-    with open(filePath, "w") as outfile :
-        json.dump(jsonObject, outfile, indent=3)
-        if (DEBUG) : print(greenText("Output to JSON success"))
